+++ src/main/java/community/whatever/onembackendjava/common/exception/notfound/NotFoundShortenUrlException.java
@@ -1,4 +1,4 @@
-package community.whatever.onembackendjava.domain;
+package community.whatever.onembackendjava.common.exception.notfound;
 
 
 public class NotFoundShortenUrlException extends RuntimeException {
