+++ src/main/java/community/whatever/onembackendjava/controller/UrlShortenController.java
@@ -1,6 +1,9 @@
-package community.whatever.onembackendjava.presentation;
+package community.whatever.onembackendjava.controller;
 
-import community.whatever.onembackendjava.application.UrlShortenService;
+import community.whatever.onembackendjava.controller.response.GetOriginUrlResponse;
+import community.whatever.onembackendjava.controller.request.ShortenUrlRequest;
+import community.whatever.onembackendjava.controller.response.ShortenUrlResponse;
+import community.whatever.onembackendjava.service.UrlShortenService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
