+++ src/main/java/community/whatever/onembackendjava/controller/request/ShortenUrlRequest.java
@@ -1,4 +1,4 @@
-package community.whatever.onembackendjava.presentation;
+package community.whatever.onembackendjava.controller.request;
 
 public class ShortenUrlRequest {
     private String originUrl;
