+++ src/main/java/community/whatever/onembackendjava/controller/response/ShortenUrlResponse.java
@@ -1,4 +1,4 @@
-package community.whatever.onembackendjava.presentation;
+package community.whatever.onembackendjava.controller.response;
 
 public class ShortenUrlResponse {
