+++ src/test/java/community/whatever/onembackendjava/application/UrlShortenServiceTest.java
@@ -3,8 +3,8 @@ package community.whatever.onembackendjava.application;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import community.whatever.onembackendjava.service.UrlShortenService;
 import java.util.NoSuchElementException;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
